MODREQMED-2 Create module skeleton - SpringBoot
